add starter components dir
