+++ starters/apps/base/index.html
@@ -9,7 +9,7 @@
       // Vite Dev Mode Only
       import '@builder.io/qwik/qwikloader.js';
       import { render, jsx } from '@builder.io/qwik';
-      import { App } from '/src/app.tsx';
+      import { App } from '/src/components/app/app.tsx';
       render(document.body, jsx(App));
     </script>
   </body>
