+++ starters/apps/base/src/typings.d.ts
@@ -4,6 +4,24 @@ declare module '*.css' {
   export default content;
 }
 
+/* Allows Typescript to import .gif files */
+declare module '*.gif' {
+  const content: string;
+  export default content;
+}
+
+/* Allows Typescript to import .jpg files */
+declare module '*.jpg' {
+  const content: string;
+  export default content;
+}
+
+/* Allows Typescript to import .png files */
+declare module '*.png' {
+  const content: string;
+  export default content;
+}
+
 /* Allows Typescript to import .svg files */
 declare module '*.svg' {
   const content: string;
