+++ starters/apps/starter-partytown/src/entry.server.tsx
@@ -8,7 +8,7 @@
 
 import { renderToString, RenderToStringOptions, QwikLoader } from '@builder.io/qwik/server';
 import { partytownSnippet } from '@builder.io/partytown/integration';
-import { App } from './app';
+import { App } from './components/app/app';
 
 /**
  * Entry point for server-side pre-rendering.
