+++ starters/apps/todo/src/components/main/main.tsx
@@ -1,0 +1,26 @@
+import { $, component, Host, onRender$ } from '@builder.io/qwik';
+import { getFilteredItems, Todos } from '../../state/state';
+import { Item } from '../item/item';
+
+/**
+ * Main body of the application which contains the list of todo items.
+ *
+ * This component only rerenders/hydrates/downloads if the list of todos changes.
+ */
+export const Main = component(
+  'main',
+  $((props: { todos: Todos }) => {
+    return onRender$(() => {
+      console.log('on:qRender => <Main/>');
+      return (
+        <Host class="main">
+          <ul class="todo-list">
+            {getFilteredItems(props.todos).map((key) => (
+              <Item item={key} todos={props.todos} />
+            ))}
+          </ul>
+        </Host>
+      );
+    });
+  })
+);
